srm: Increase default database queue
When receiving bulk requests, such as bulk stage requests, we temporarily
insert a large number of database updates in the queue. Thus the queue limit
should be large enough to encompass these bulk requests to avoid overflow.

This patch increases the default to 10000 as 1000 is usually too low for
our sites.

This patch is only for 2.10 as 2.11 introduced a throttling mechanism.

Target: 2.10
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/7980/