pool: Fix repository statistiscs collection
The pool precomputes the 'rep ls -s' output. The statistics collector has a bug
in that it disregards the events reported in the StateChangeEvent. It only looks
at the state in the CacheEntry. This means that during pool startup it will
not react to the state change from NEW and thus none of the files that existed
when the pool started are included in the statistics.

The patch also rewrites a related if condition to make its intend easier to
undertand.

Target: trunk
Request: 2.12
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/7999/