cells: Prevent removal of wrong route
When a domain disconnects, the tunnel removes the routes it added when the
domain connected. The routing table logic fails to check whether the route
removed is actually the same route - it removes any route that matches the
key.

The consequence is that if you got two domains with the same well known cell,
the first one to connect is correctly registered, and the second one will
- correctly - not be registered. However if the second domain disconnects, the
tunnel will remove the route for the well known cell in the first domain.

The fix modifies a failure case in the LocationMgrTunnel. That case should only
be relevant in case of bugs or in the pressence of manual modifications to the
routing table. In the former case, logging is probably a better solution. In
the latter case, respecting the manual route is probably a better solution too.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8022/