all: Simplify handling of cell routing failures
When sending a message, there were two ways that the sender could receive a
NoRouteToCellException: The first was when the failure happened on the first
hop. In that case the send message would throw the NoRouteToCellException. If
the failure happened later in the process, the NoRouteToCellException would be
returned as a regular reply. For sendAndWait it would be rethrown as a local
exception. Otherwise, if a callback had been registered, it would be called
with the exception. Finally, if none of the above applied, the exception would
be delievered to a message handler via messageArrived.

A correct cell had thus to implement both an exception handler, and a callback
and/or message handler to correctly deal with message delivery errors. Worse,
one may be mislead to believe that simply catching the exception on send is
enough to handle message delivery errors (after all, it is declared to throw
NoRouteToCellException). Finally, in most cases we simply logged the error and
carried on, which means that there was a lot of boilerplate from being forced
to handle the NoRouteToCellException on send.

This patch removes one of the above options: Throwing an exception on send.
Instead, NoRouteToCellException is always reported as a regular reply, either
to be rethrown by sendAndWait, or to be repoted to a callback or a message
handler. Sub-classes of AbstractCell have a message handler, that log such
errors (I am lazy - UniversalSpringCell should not log this itself - instead
all Spring cells should provide a custom handler; I may opt to do that later).

This naturally touches a large part of the code base, although in most cases it
is just about removing exception handlers. In some cases the exception handlers
caught Exception rather than NoRouteToCellException. In those cases I changed
the handler to catch RuntimeException to avoid changing other aspets of the
logic. Most of these cases should probably let the RuntimeException propagate,
but that can be done later.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8037/