nfs: share mover for the same client
the linux client have following bahavoiur:

on a client host we have two processes - A and B.
Process A opens a file and reads some data:

OPEN -> LAYOUTGET -> READ.

Process B opens the same file:

OPEN -> READ.

Client B obviously re-used layout information, but accesses
pool with a different state id (from the second open).

The nfs spec assumes, that stateids globally visible and valid,
but this is not the case with the dcache.

With this change we try to detect such cases and share
mover with two opens, if they comes from the same client.

Acked-by: Gerd Behrmann
Ticket: #8691
Target: master
Requires-notes: no
Requires-book: no