cells: Simplify routing code
Keeping track of when destCore, cellName and domainName where updated made the
code hard to read. The patch eliminates cellName and domainName. Since
CellAddressCore is immutable, we can leave it to the JVM to optimize this if
needed.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8041/