cells: Fix route loop detection
Cells limits the number of times it consults a routing table. This is to prevent
rewriting loops. Unfortunately the implementation is broken as the test is
placed inside the big for loop and the iteration variable never reaches the
limit. Instead the loop is left and the message is silently discarded.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8043/