cells: Detect end of destination path
When cells tries to find a route for a path, it doesn't detect when it reaches
the end of the path - in particular if the last element is a domain address
(the cell component is an asterix), it would just keep looping until the
maximum number of iterations allowed was reached.

This patch adds a check for the end of path and returns a no route to cell
exception to the sender when the end is reached.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8045/