cells: Refactor routing loop
The routing loop disitinguishes between the first and subsequent iterations,
making the logic complicated. This patch refactor the code by unfolding the
first iteration. This greatly simplifies the logic.

The patch introduces one behaviour change to fix a problem that became obvious
once the code was refactored: If sending a message to foo@local with deliver
locally set to true and deliver remotely set to false, and foo not being a
local cell, the old code would attempt remote delivery anyway as the check for
'!remotely' was done in an else branch for the 'locally' check.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8063/