cells: Only use RoutedMessageEvents when sending to routers
CellGlue delivers a message as either a MessageEvent or a RoutedMessageEvent.
The difference is that the former gets decoded in CellNucleus while the latter
is delivered in encoded form as it is intended to be routed to another domain.

Cells has the concept of alias routes. These do not merely determine how to
deliver the message; alias routes actively rewrite the destination path, thus
the routing table can dictate not only that a message is delivered *through* a
local cell, it can also dictate that a message is deliverd *to* a local cell.
Yet CellGlue always delivered such messages as a RoutedMessageEvent.

CellNucleus contained a hack that would detect if the cell is a CellTunnel and
then rewrite the such events to regular MessageEvent. Not only was this
unnecessarily complicated, it also meant that an alias cell could not rewrite
to a tunnel cell, as the tunnel cell would have still received a
RoutedMessageEvent and passed the message on to the other domain.

This patch resolves this by letting CellGlue detect whether the message is
addressed to the cell to which the message is about to be delievered and
use the proper event. This also means that the unrouted and routed code
paths are ones again sufficiently similar that the loop unfolding done in
one of the earlier patches is no longer necessary.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8066/