cells: Move message encoding/decoding to CellGlue
A previous patch moved the encoding to CellNucleus as that's where the decoding
was done. This patch moves both the encoding as well as the decoding back to
CellGlue. This achieves all the same benefits of high coherence, while
providing a single place to encode and decode, thus making the code more DRY.

Further, CellGlue already contained code to decide whether to decode or
not (i.e. whether to use RoutedMessageEvent or MessageEvent).

As a side effect, the patch changes the decoding point to be before the message
is queued in the receiving cell. Also, in case of deserialization errors, an
NoRouteToCellException is returned to the sender.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8068/