Restore compile time compatibility with pcells
pcells depends on a number of dCache classes. This patch restores compile time
compatibility with pcells and moves the required classes to dcache-vehicles.
When paired with a number of fixes to pcells, this allows pcells to avoid the
majority of the dCache dependencies.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Christian Bernardt <christian.bernardt@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8069/