cells: Drop digest implementations
We use those provided by the JRE and bouncy castle, so there is no need
to have our own implemetations.

Moves the crypt hash to the login service package (it can eventually be
replaced with the implementation in bouncy castle, but that's not done in
this patch).

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8082/