pool-selection-unit: refactor commands to delegate to a provider
In order to be able to notify the replica system of
changes to the selection unit made through
the admin commands, the following refactoring has been
done:

1.  Defined an interface, PoolSelectionUnitAccessor, which
    abstracts the methods which correspond to the CLI calls,
    but decoupled from the Args parameter type.

2.  Moved the CLI calls to a separate class which composes
    an instance of the PoolSelectionUnitAccessor and
    passes in the correct parameters to the defined methods.

3.  Made PoolSelectionUnitV2 implement both PoolSelectionUnit
    and PoolSelectionUnitAccess.

4.  Created PoolSelectionUnitDecorator which implements
    PoolSelectionUnitAccess and also composes an instance
    of the same to use as delegate.

5.  Added the new commands class as a bean to the poolmanager.xml.
    Defined a new plugin property for the implementer of
    the new access interface.

The decorator also implements the CellSetupProvider methods
to be able to deactivate the interception of the commands on
load/startup.

This patch does not attempt to do any other refactoring of
the PoolSelectionUnitV2 code, nor does it replace the
older style commands with the newer.

Target: master
Acked-by: Tigran