httpd: Fix ever increasing source address path
A recent change touched how we send messages and how we add the sender
to CellMessage. That change assumed that CellMessages would not be sent
multiple times (with the exception of replying or forwarding it). The
httpd service breaks this assumption, causing the source address to be
added over and over.

Target: trunk
Require-notes: no
Require-book: no