cells: Allow route duplication by key
There are use cases for allowing more than one route for a given routing
target. In particular in fail over or switch over scenarios in which a new well
known cell or domain is brought up before the old one is taken down. Before,
the new route would be rejected if the previous cell was still known. Thus one
had to kill the old cell/domain first and then bring up the new one, leaving a
window in which the service was absent.

With this patch, the new route can be installed while the existing cell is
still running. The new route will be placed after the existing route, which
means it will not be used. Once the old cell or domain disconnects, the new
route becomes the first matching route and is used immediately.

The patch also alters the lookup order of alias routes vs wellknown routes.
Before, a wellknown route had precedence over an alias route, meaning it was
impossible to rewrite an address if a wellknown route existed. This is now
fixed. A duplicate route lookup for an exact/alias route is also removed.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8091/