doors: Clean up mover timeout interpretation
Several doors have the concept of a mover timeout - some have it configurable,
others have it hard coded. There is however some confusion about what exactly
this timeout covers: The documented behaviour is that it defines the time a
mover can be queued for execution on the pool. The way it is implemented in the
Transfer class though, the mover timeout is also used to define how long to
give the pool to create the mover in the first place. Those two operations do
however have widely different timebounds.  Creation a mover on a well
performing pool should be a sub-second operation, while the time a mover stays
in the queue may be unbounded.

This patch separates these concepts. The mover creation timeout is now defined
by the general pool stub timeout, while the mover queueing timeout is defined
as before. This actually restored the behaviour from before
Transfer#selectPoolAndStartMover was introduced.

The patch also reintroduces a timeout for pool selection, again restoring the
behaviour we had before selectPoolAndStartMover was introduced. The timeout
defined for the pool manager stub (per the *.service.poolmanager.timeout
settings) defines a timeout for the pool selection independent of the retry
policy.

Both pool selection and mover creation are however subject to the overall
timeout of preparing a transfer as defined in the transfer retry policy. How
this timeout is defined is door dependent.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8101/