info/info-provider: publish valid IPv6 addresses
An IPv6 address within a URL must be escaped by being placed in square
brackets; an IPv4 address and a hostname must not be so escaped.
Currently info-provider publishes door interfaces without any escaping,
which is broken for IPv6 addresses.

Rather than placing the logic for escaping IPv6 address in
info-provider, info service is updated to publish an additional metric
('url-name') which is the interface's name in a format suitable for
building a URL.

The info-provider is updated to use this metric.  For backwards
compatibility, if the 'url-name' metric is missing then the 'FQDN'
metric is used as a fall-back.

Target: master
Patch: https://rb.dcache.org/r/8108/
Acked-by: Gerd Behrmann
Request: 2.12
Request: 2.11
Request: 2.10
Requires-book: no
Requires-notes: yes