cleaner,pnfsmanager,replica,spacemanager: Rework delete notification
Delete notifications are essential for both replica manager and space manager,
yet these are lossy notifications send through the broadcast service with no
retry on failures. Further more, the current solution relies on a combination
of notification on name space entry deletion and cleaner notification - the
latter because for space manager we don't want to release space until the
physical files have been deleted, the former because the latter is only sent
when pool locations have been registered as cleaner is otherwise not triggered.

This patch reworks these to a single notification sent when all physical disk
and tape locations have been cleaned - even when there were no locations at
all. This is implemented by extending the trash trigger in chimera to insert an
extra record with the new itype 2. Cleaner will send a message to configured
services when all other trash entries for a file have been cleaned. The itype 2
entry is not removed until all configured services have returned a positive
response.

Besides modifying the trigger in chimera, the patch also adds a combined index
on the itype and ilocation fields for the trash table.

The patch deprecates some properties and introduces other. Most notably, it is
now essential that *.enable.space-reservation and *.enable.replica are
configured for cleaner! Otherwise cleaner will not send the notification
message to the respective services.

Question to reviewer: Would it make sense to alter the default of
dcache.enable.space-reservation to true? Currently, the effect of adding a
spacemanager service but forgetting to enable dcache.enable.space-reservation
on some services is severe (dCache may silently do the wrong thing); the
opposite (forgetting to disable it when not using the spacemanager service)
would simply result in a non-functioning installation. Thus is seems 'true' is
the safer default.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8111/