admin: Avoid line breaks when not connection from a TTY
The legacy shell is used both interactively and from scripts. After the upgrade
to jline 2, the 80 character default terminal width use when the client did not
allocate a pseudo TTY causes unintended line feeds to be inserted into the output.

This patch fixes this problem and also disables the use of any ANSI sequences
when the client does not provide a TTY.

It was tempting to define the terminal for a non-tty session as not supported
in jline, causing it to disable most of its features, but this would also have
suppressed echoing the commands to the output. Although that would have been
the correct thing to do when using the shell in a pipe, we have always echoed
the command and it is likely that existing scripts rely on this.

Target: trunk
Request: 2.12
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8122/