spacemanager: Set default of dcache.enable.space-reservation to true
When selecting default values, it is important to consider the consequences of
a site forgetting to configure a particular property in case the default isn't
suitable.

In a dCache instance that does use space manager, forgetting to set
dcache.enable.space-reservation to true will lead to silent corruption as space
manager is not notified about file deletion and file flushing. On the other
hand, in a dCache without space reservation, forgetting to set
dcache.enable.space-reservation to false will merely result in a non-functional
dCache - one that will loudly protest about the lack of a space manager
service.

Setting dcache.enable.space-reservation to true is thus the safer default. A
site not using space reservations can either choose to change the value of the
property or to add a space manager service. The latter used to not be an option
as enabling space manager forced all uploads to link groups, but that's not
true these days; nowadays you can add a spacemanager service without ever
creating a link group and everything will continue to work just fine.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8123/