cells: Drop routes and event listeners on cell shutdown
This patch ensures that specific routes to a cell are removed when
the cell shuts down. The patch also ensures that event listeners
registered for a particular cell are deregistered when the cell
shuts down.

Finally, the patch fixes a race condition in the cell event listener
registration.

In the existing code, the bugs fixed by this patch are unlikely to
cause problems and thus I am not requesting for the patch to be
backported.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8134/