cells: More protection against cyclic routing
The patch adds a check to avoid sending back a message to the domain we just
got it from, unless we somehow updated the destination. E.g.  if you (for
whatever reason) have a wellknown route in dCacheDomain pointing too fooDomain,
but fooDomain doesn't contain the target cell, this patch prevents that the
message it sent back to dCacheDomain (which would just send it back to
fooDomain again).

This is not protection against every kind of cyclic routes, as there are
legitimate cycles involving alias rewriting routes that we cannot distinguish
from true cycles. The real motivation for this patch is an upcomming change I
have in the pipeline. That patch will rely on messages not being sent back to
where we got them from.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8139/