doors: Log real path in billing
When introducing the upload directories (i.e. unique write TURLs), I choose to
log the upload path in billing and only log the real path in the billing entry
created by the SRM. The idea was to allow us to distinguish between uploads
belonging to different SRM sessions.

Nobody seems to appreciate this use of the path in the billing and at the same
time the use of the upload path causes lots of confusion. This patch thus
ensures that the real path is logged to billing instead. This also makes sense
when viewing it from the point of this being an accounting record: If you want
to hold the user responsible, the record should be phrased in terms the user
understands (i.e. the real path).

The real path is only available as the "path" flag in StorageInfo, so this
patch extracts that field and injects it into the relevant mesages. To make it
clear that this is a path intended for billing, a number of setters and getters
of various messages have been renamed (the fields in the messages have not been
renamed to maintain backwards compatibility).

The temporary upload path is available as a new field in billing called
transferPath. It is not used by default, but sites that want the temporary path
logged can do so.

One unused field from the XrootdProtocolInfo has been removed.

Target: trunk
Request: 2.12
Request: 2.11
Request: 2.10
Require-notes: yes
Require-book: yes
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8140/