cells: Introduce publish-subscribe messaging
This patch introduces an alternative to the broadcast service. It allows cells
to subscribe to named topics and any message sent to that topic by other cells
are delivered to every cell that has subscribed to it. This addresses similar
usecases as the broadcast service, but without an explicit service and with a
much more flexible mechanism to subscribe to topics.

Like all other cell message delivery, publish-subscribe messaging is best
effort.  The system is deliberately kept simple like the rest of cell
messaging. At the heart of the feature is the new topic-route. These facilitate
routing of messages to cells and domains and in contrast to other routes,
delivery through these routes duplicate the message. Thus the route doesn't
cosume the message, allowing it to be delivered to more than one recepient.

Like for the rest of cell messaging, RoutingManager takes care of adverticing
topic subscriptions along default routes.

Topics are cell addresses and thus share the same name space as cell names. I
propose prefixing all topics with "Topic" as a naming convention. This overlap
in name space is deliberate, both to keep things simple and because it allows
some use cases that would otherwise be difficult. From the point of view of the
sender, there is no difference between sending a message to a topic or to
another cell.

As a proof of concept, the patch updates the delivery of pool-up notifications
to use the PoolUpTopic. Pool manager and cleaner are updated to subscribe to
this topic. Some unused parts of the broadcast service are removed. If
accepted, the remaining uses of the broadcast service will be updated too and
the broacast service will be removed.

Introduces: cleaner.cell.subscribe and poolmanager.cell.subscribe.

Obsoletes: cleaner.service.broadcast

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8144/