srm: Fix loading of jobs during restart
In 2.12 we changed the semantics of
srm.persistence.enable.store-transient-state.  As a consequence,
FinalStateOnlyJobStorageDecorator may store jobs in non-final states if the
update is forced. Thus the implementation of getActiveJobs() is wrong, since
there may be non-final jobs in the database. The observable symptom was that
with our default configuration, jobs would not be loaded back after a restart.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.12
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8154/