srm: Force save when job becomes RQUEUED
The intention was that jobs that are in RQUEUED or READY can survive an SRM
restart, however that doesn't work if we don't save that state to the database.
This job achieves this by forcing the save when the job becomes RQUEUED. We
don't force save it again when it becomes READY as we treat those states the
same on restart.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.12
Request: 2.11
Request: 2.10
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8156/