srmclient: consider copyjobfile when deciding whether to delegate
By default, when the srmcp command is invoked it chooses whether or
not to delegate the credential based on if the transfer is a
second-party (upload/download) or a third-party transfer.

The current implementation does not consider the copyjobfile which
allows the user to specify multiple files to be copied in one
go.  The result is that no delegation happens, which results in
copyjobfile-based requests to fail if they involve third-party
transfers.

This patch adds support for checking whether the copyjobfile contains
any third-party transfers.  The srmcp client will delegate by
default if the copyjobfile contains third-party transfers and not
otherwise.

Target: master
Ticket: http://rt.dcache.org/Ticket/Display.html?id=8716
Patch: https://rb.dcache.org/r/8157/
Acked-by: Gerd Behrmann
Request: 2.12