admin: Avoid endless exception loop upon client disconnect
If the client disconnects such that the admin service fails to write to the
output stream, the services enters an endless loop in which Apach SSHD throws
an SSHException (subclass of IOException). The admin service catches this and
attempts to report an error to the client, which will trigger another
exception, repeating this process indefinitly.

This patch makes sure that IOExceptions are propagated.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.12
Acked-by: Albert Rossi <arossi@fnal.gov>
Patch: https://rb.dcache.org/r/8186/