dcache: Replace loginbroker service by publish-subscribe topic
This patch removes the loginrboker service. Instead a publish-subscribe topic
is introduced. All doors publish to this topic and all interested parties
subscribe to it.

The information published is still referred to as login broker information
and the admin shell commands are prefixed by 'lb' for familarity. The service
itself is however gone.

Doors may now be tagged. Consumers may filter by these tags. This provides a
replacement for the common use case of partitioning doors by using multiple
login brokers. By default, all doors are tagged by 'srm' and 'glue' to indicate
that they may be used by the srm and be published into the glue schemas (the
latter is not implemented - I hope Paul can update the info provider).

The login broker information is also extended with information about authorized
read and write paths. This provides a generalized way to define read-only
doors - the xrootd door makes use of this. The SRM is updated to take this
information into account when selecting doors.

A second publis-subscribe topic allows for explicit updates to be requested
from all doors. Doors subscribe to this second topic. Several cells use this
to quickly discover all doors when starting up.

Both publishers and subscribers detect no-route-to-cell exception to discover
if their domain is likely not connected to the rest of dCache. In that case
we watch out for the addition of topic and default routes and resubmit when
these become available.

To support pcells, the PcellsCommand mimics the old login broker and collects
updates from doors.

Webadmin as a webapp doesn't have a good way to subscribe to the topic. Until
a better solution is found, an ugly hack is used to inject the subscribing
bean through the domain context.

All *.service.loginbroker.* properties are obsolete. New *.loginbroker.*
properties have been introduced for some obsolete properties.

WebDAV doors both publish and subscribe to login broker updates. This is
because WebDAV doors need to discover SRM doors as these provide delegation
endpoints. This should be replaced by a dedicated topic for publishing such
endpoints. This would decouple WebDAV from SRM (there is nothing specific
about SRM that says we must couple this service to the SRM) and it would
avoid publishing information about all doors to all WebDAV doors (for a
deployment with many doors, this could become quite a lot of updates).

Upon upgrade, loginbroker services must be removed from the layout file. If
more than one loginbroker was used, the tags must be set to support that
usecase.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8206/