cells: Take additivity into account in logback cells filter
We support per cell log levels. This is implemented using a LogBack turbo
filter. Such a filter is invoked very early in the logging process and
avoids most of the logging overhead in case the log target is disabled.

Our filter however failed to take additivity into account. A logger bound to
one appender and which isn't additive will only log to that one appender, yet
our filter considered the appenders attached to ancestors of the logger too.
The log message was not actually logged, as the filter decision for a
particular logger and a particular appender was correct - it was merely the
check whether there is any enabled appender for a log target that was wrong.
Thus this bug was merely a performance problem as early checks for whether
a log event should be logged had false positives.

In particular this speeds up cell to cell communication. Single domain cell to
cell communication is as much as 40% faster (using a trivial ping message -
the relative speedup will be slower with more complex messages).

The pach also eliminates some wrapping of logger names and expensive string
comparisons on hash map lookups.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8213/